initial commit
Develop an application that repeatedly reads numbers until the user enters "done". once "done" is entered, print out the total, count,average of the numbers.if the user enters anything other than 
a number, detect their mistake using try and except and print an error message and skip to the next nuumber.