initial commit
Develop an application to open the file ( say.romeo.txt ) and read it line by line . for each line, split the line into words using the split function.
for each word ,check to see if the word is already in a list. if the word is not in the list, add it to the list. when the program completes, sort and print the resulting
words in alphabetical order .